Refactored user view register.Created a custom user creation form and added regex to check mitigate against unwanted user input.

+++ entries/views.py
@@ -79,5 +79,8 @@ def edit_entry(request, entry_id):
 @login_required(login_url="/login/")
 def delete_entry(request, entry_id):
     entry = get_object_or_404(Entry, id=entry_id)
-    entry.delete()
-    return redirect('get_entries')
+    if request.user != entry.owner:
+        return redirect('get_entries')
+    else:
+        entry.delete()
+        return redirect('get_entries')
